Combine Add / Remove Treasury NFT collection actions
