+++ packages/i18n/locales/en/translation.json
@@ -437,7 +437,9 @@
     "logo": "Logo",
     "majority": "Majority",
     "manageMembersActionDescription": "Add, update, or remove members from the DAO.",
+    "manageStorageItemsDescription": "Manage Storage Items associated with your DAO. Set or remove key / value pairs.",
     "manageSubDaosActionDescription": "Recognize or remove SubDAOs from the DAO.",
+    "manageTreasuryNftsDescription": "Manage NFT Collections displayed on your DAO's treasury page.",
     "members": "members",
     "membershipDaoNotMemberInfo_dao": "You are not a member of {{daoName}}. To become a member, existing members must vote to add you.",
     "membershipDaoNotMemberInfo_proposal": "You were not a member of {{daoName}} when this proposal was created, so you cannot vote on it.",
@@ -702,6 +704,7 @@
     "manageMembers": "Manage members",
     "manageStaking": "Manage staking",
     "manageSubDaos": "Manage SubDAOs",
+    "manageTreasuryNfts": "Manage Treasury NFTs",
     "memberOfDao": "Member of {{daoName}}",
     "members": "Members",
     "membership": "Membership",
