+++ packages/stateful/actions/components/index.tsx
@@ -1,15 +1,14 @@
+export * from './ActionCard'
+export * from './ActionsRenderer'
 export * from './AddCw20'
-export * from './AddCw721'
 export * from './Custom'
 export * from './Execute'
 export * from './Instantiate'
+export * from './ManageCw721'
+export * from './MigrateContract'
+export * from './NativeCoinSelector'
 export * from './RemoveCw20'
-export * from './RemoveCw721'
 export * from './Spend'
 export * from './StakingActions'
-export * from './NativeCoinSelector'
-export * from './ActionCard'
-export * from './ActionsRenderer'
-export * from './UpdateInfo'
-export * from './MigrateContract'
 export * from './UpdateAdmin'
+export * from './UpdateInfo'
