+++ packages/types/actions.ts
@@ -13,8 +13,7 @@ export enum CoreActionKey {
   StakingActions = 'stakingActions',
   AddCw20 = 'addCw20',
   RemoveCw20 = 'removeCw20',
-  AddCw721 = 'addCw721',
-  RemoveCw721 = 'removeCw721',
+  ManageCw721 = 'manageCw721',
   TransferNft = 'transferNft',
   MintNft = 'mintNft',
   BurnNft = 'burnNft',
